fixed stuff

+++ database/migrations/2021_04_15_171526_create_lesson_grades_table.php
@@ -17,7 +17,7 @@ class CreateLessonGradesTable extends Migration
             $table->id();
             $table->foreignId('lesson_id')->constrained();
             $table->foreignId('enrollment_id')->constrained();
-            $table->integer('grade');
+            $table->float('grade');
             $table->timestamps();
 
         });
