+++ database/migrations/2021_04_15_171542_create_exam_grades_table.php
@@ -16,7 +16,7 @@ class CreateExamGradesTable extends Migration
         Schema::create('exam_grades', function (Blueprint $table) {
             $table->id();
             $table->foreignId('enrollment_id')->constrained();
-            $table->integer('grade');
+            $table->float('grade');
             $table->timestamps();
         });
     }
